% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomRegMethModel.R
\name{binomRegMethModelSummary}
\alias{binomRegMethModelSummary}
\title{binomRegMethModelSummary Lorem ipsum dolor sit amet}
\usage{
binomRegMethModelSummary(GamObj, var.cov.alpha, new.par, edf.out, edf1.out,
  X_d, resi_df, Quasi, scale, RanEff, re.test, Z)
}
\arguments{
\item{GamObj}{Lorem ipsum dolor sit amet}

\item{var.cov.alpha}{Lorem ipsum dolor sit amet}

\item{new.par}{Lorem ipsum dolor sit amet}

\item{edf.out}{Lorem ipsum dolor sit amet}

\item{edf1.out}{Lorem ipsum dolor sit amet}

\item{X_d}{Lorem ipsum dolor sit amet}

\item{resi_df}{Lorem ipsum dolor sit amet}

\item{Quasi}{Lorem ipsum dolor sit amet}

\item{scale}{Lorem ipsum dolor sit amet}

\item{RanEff}{Lorem ipsum dolor sit amet}

\item{re.test}{Lorem ipsum dolor sit amet}

\item{Z}{Lorem ipsum dolor sit amet}
}
\value{
Lorem ipsum dolor sit amet
}
\description{
Lorem ipsum dolor sit amet

Lorem ipsum dolor sit amet
}
\author{
Kaiqiong Zhao
}
