% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSMethEM.R
\name{BSMethEM}
\alias{BSMethEM}
\title{A smoothed-EM algorithm to estimate covariate effects and test regional association in Bisulfite Sequencing-derived methylation data}
\usage{
BSMethEM(data, n.k, p0 = 0.003, p1 = 0.9, Quasi = TRUE,
  epsilon = 10^(-6), epsilon.lambda = 10^(-3), maxStep = 200,
  detail = FALSE, binom.link = "logit", method = "REML",
  covs = NULL)
}
\arguments{
\item{data}{a data frame with rows as individual CpGs appeared in all the samples. The first 4 columns should contain the information of `Meth_Counts` (methylated counts), `Total_Counts` (read depths), `Position` (Genomic position for the CpG site) and `ID` (sample ID). The covariate information, such as disease status or cell type composition are listed in column 5 and onwards.}

\item{n.k}{a vector of basis dimensions for the intercept and individual covariates. \code{n.k} specifies an upper limit of the degrees of each functional parameters.}

\item{p0}{the probability of observing a methylated read when the underlying true status is unmethylated.}

\item{p1}{the probability of observing a methylated read when the underlying true status is methylated.}

\item{Quasi}{whether a Quasi-likelihood estimation approach will be used}

\item{epsilon}{numeric; stopping criterion for the closeness of estimates of spline coefficients from two consecutive iterations.}

\item{epsilon.lambda}{numeric; stopping criterion for the closeness of estimates of smoothing parameter \code{lambda} from two consecutive iterations.}

\item{maxStep}{the algorithm will step if the iteration steps exceed \code{maxStep}}

\item{detail}{indicate whether print the number of iterations}

\item{binom.link}{the link function used in the binomial regression model; the default is the logit link}

\item{method}{the method used to estimate the smoothing parameters. The default is the "REML" method which is generally better than prediction based criterion \code{GCV.cp}}

\item{covs}{a vector of covariate names. The covariates with names in \code{covs} will be included in the model and their covariate effects will be estimated.}
}
\value{
This function return a \code{list} including objects:
\itemize{
\item \code{est}: estimates of the spline basis coefficients alphas
\item \code{lambda}: estimates of the smoothing parameters for each functional paramters
\item \code{est.pi}: predicted methylation lelves for each row in the input \code{data}
\item \code{ite.points}: estimates of \code{est}, \code{lambda} at each EM iteration
\item \code{cov1}: estimated variance-covariance matrix of the basis coefficients alphs
\item \code{reg.out}: output from the reginal zero effect test
\item \code{chi.sq}: chi-square statistics for each covariates (including intercept) w.r.t the zero regional effect tests
\item \code{pvalue}: the p value from the regional tests for each covariate
\item \code{pvalue.log}: the log of \code{pvalue}
\item \code{SE.out} a matrix of the estimated pointwise Standard Errors (SE); number of rows are the number of unique CpG sites in the input data and the number of columns equal to the total number of covariates fitted in the model (the first one is the intercept)
\item \code{SE.pos} the genomic postions for each row of CpG sites in the matrix \code{SE.out}
\item \code{Beta.out} a matrix of the estimated covariate effects beta(t), here t denots the genomic positions.
\item \code{ncovs} number of functional paramters in the model (including the intercept)
}
}
\description{
This function fits a binomial regression model where the outcome - methylated reads- are contaminated by known error rates \code{p0} and \code{p1} and the covariate effects are smoothly varying across genomic positions. The functional parameters of the smooth covariate effects are first represented by a linear combination of a bunch of restricted cubic splines (with dimention \code{n.k}), and a smoothness penalization term which depends on the smoothing parameters \code{lambdas} is also added to control smoothness.

The estimation is performed by an iterated EM algorithm. Each M step constitutes an outer Newton's iteration to estimate smoothing parameters \code{lambdas} and an inner P-IRLS iteration to estimate spline coefficients \code{alpha} for the covariate effects. Currently, the computation in the M step depends the implementation of \code{gam()} in package \code{mgcv}.
}
\examples{
#------------------------------------------------------------#
data(RAdat)
head(RAdat)
RAdat.f <- na.omit(RAdat[RAdat$Total_Counts != 0,])
out <- BSMethEM(data=RAdat.f, n.k = rep(5,3), p0 = 0.003307034, p1 = 0.9,
epsilon = 10^(-6), epsilon.lambda = 10^(-3), maxStep = 200, detail=FALSE)
}
\seealso{
\link[mgcv]{gam}
}
\author{
Kaiqiong Zhao
}
