% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSMethEMUpdate.R
\name{BSMethEMUpdate}
\alias{BSMethEMUpdate}
\title{One-step update inside the EM algorithm for fitting the functional parameter theta.0 and beta.}
\usage{
BSMethEMUpdate(data, pi.ij, p0, p1, n.k, binom.link, method, Z,
  my.covar.fm, Quasi = T, scale)
}
\arguments{
\item{data}{a matrix with  \code{n*p} rows and the columns should include Y, X, Posit, Ctype and ID}

\item{pi.ij}{an initial value (or value from last step) for the fitted probability \code{pi_ij} for each CpG site of each individual}

\item{p0}{the probability of observing a methylated read when the underlying true status is unmethylated.}

\item{p1}{the probability of observing a methylated read when the underlying true status is methylated.}

\item{n.k}{a vector of basis dimensions for the intercept and individual covariates. \code{n.k} specifies an upper limit of the degrees of each functional parameters.}

\item{binom.link}{the link function used in the binomial regression model; the default is the logit link}

\item{method}{the method used to estimate the smoothing parameters. The default is the "REML" method which is generally better than prediction based criterion \code{GCV.cp}}

\item{Z}{the covariate matrix used in BSMethEM}

\item{my.covar.fm}{the formula fitted in the GAM}
}
\value{
The function returns a list of following objects
\itemize{
\item \code{pi.ij}  fitted methylation proportions
\item \code{par} updated basis coefficients
\item \code{lambda} updated smoothing parameters
\item \code{edf1} Effective degree of freedoms for each smooth terms in the model}
}
\description{
Since the presence of conversion errors, the true methylated counts at a CpG sites are unknown. The first step inside this function is the E-step, calculating the conditional expectation of methylated counts given the observed counts. Then the second step involves maximizing the penalized likelihood using the avaialbe function \code{gam} in Package \code{mgcv}. This function fits the model where there is no random effect for individuals. Thus, the log-likelihood are sum of two parts; one is the normal approximated log-likelihood for binomial outcomes and the second part quantifies the penalization of the wiggliness of two functional parameters theta.0 and beta.
}
\examples{
#------------------------------------------------------------#
}
\author{
Kaiqiong Zhao
}
