% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomRegMethModel.R
\name{hessianComp}
\alias{hessianComp}
\title{hessianComp Lorem ipsum dolor sit amet}
\usage{
hessianComp(w_ij, new.par, new.lambda, X, Y, my.design.matrix, gam.int, Z,
  pred.pi, p0, p1, disp_est, RanEff, N)
}
\arguments{
\item{w_ij}{Lorem ipsum dolor sit amet}

\item{new.par}{Lorem ipsum dolor sit amet}

\item{new.lambda}{Lorem ipsum dolor sit amet}

\item{X}{Lorem ipsum dolor sit amet}

\item{Y}{Lorem ipsum dolor sit amet}

\item{my.design.matrix}{Lorem ipsum dolor sit amet}

\item{gam.int}{Lorem ipsum dolor sit amet}

\item{Z}{Lorem ipsum dolor sit amet}

\item{pred.pi}{Lorem ipsum dolor sit amet}

\item{p0}{Lorem ipsum dolor sit amet}

\item{p1}{Lorem ipsum dolor sit amet}

\item{disp_est}{Lorem ipsum dolor sit amet}

\item{RanEff}{Lorem ipsum dolor sit amet}

\item{N}{Lorem ipsum dolor sit amet}
}
\value{
Lorem ipsum dolor sit amet
}
\description{
Lorem ipsum dolor sit amet

Lorem ipsum dolor sit amet
}
\author{
Kaiqiong Zhao
}
