% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_BSMethEM.R
\name{plot_BSMethEM}
\alias{plot_BSMethEM}
\title{A smoothed-EM algorithm to estimate covariate effects and test regional association in Bisulfite Sequencing-derived methylation data}
\usage{
plot_BSMethEM(BEM.obj, mfrow = NULL, same.range = F)
}
\arguments{
\item{BEM.obj}{an output object from function \code{BSMethEM}}

\item{mfrow}{the plot parameters to specify the layout of each plot}

\item{same.range}{specify whether the plots should be in the same vertical scale}
}
\value{
This function return a plot of smooth covariate effects and its pointwise confidence intervals
}
\description{
This function fits a binomial regression model where the outcome - methylated reads- are contaminated by known error rates \code{p0} and \code{p1} and the covariate effects are smoothly varying across genomic positions. The functional parameters of the smooth covariate effects are first represented by a linear combination of a bunch of restricted cubic splines (with dimention \code{n.k}), and a smoothness penalization term which depends on the smoothing parameters \code{lambdas} is also added to control smoothness.

The estimation is performed by an iterated EM algorithm. Each M step constitutes an outer Newton's iteration to estimate smoothing parameters \code{lambdas} and an inner P-IRLS iteration to estimate spline coefficients \code{alpha} for the covariate effects. Currently, the computation in the M step depends the implementation of \code{gam()} in package \code{mgcv}.
}
\examples{
#------------------------------------------------------------#
head(RAdat)
RAdat.f <- na.omit(RAdat[RAdat$Total_Counts != 0,])
out <- BSMethEM(data=RAdat.f, n.k = rep(5,3), p0 = 0.003307034, p1 = 0.9,
epsilon = 10^(-6), epsilon.lambda = 10^(-3), maxStep = 200, detail=F)
plot_BSMethEM(out, same.range = F)
}
\author{
Kaiqiong Zhao
}
