% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAdat2.R
\docType{data}
\name{RAdat2}
\alias{RAdat2}
\title{Lorem ipsum dolor sit amet}
\format{
A data frame of 6064 rows and 13 columns. Each row represents a CpG site for a sample. Columns include in order
\describe{
\item{Meth_Counts}{Number of methylated reads}
\item{Total_Counts}{Total number of reads; read-depth}
\item{Position}{Genomic position (in bp) for the CpG site}
\item{ID}{indicates which sample the CpG site belongs to}
\item{ACPA4}{Lorem ipsum dolor sit amet}
\item{Age}{Lorem ipsum dolor sit amet}
\item{Sex}{Lorem ipsum dolor sit amet}
\item{Smoking}{Lorem ipsum dolor sit amet}
\item{Smoking_NA}{Lorem ipsum dolor sit amet}
\item{PC1}{Lorem ipsum dolor sit amet}
\item{PC2}{Lorem ipsum dolor sit amet}
\item{PC3}{Lorem ipsum dolor sit amet}
\item{PC4}{Lorem ipsum dolor sit amet}
}
}
\source{
Lorem ipsum dolor sit amet
}
\usage{
RAdat2
}
\description{
This example data include methylation levels on a region with 208 CpGs for 116 blood samples.
}
\keyword{datasets}
