% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomRegMethModelPlot.R
\name{binomRegMethModelPlot}
\alias{binomRegMethModelPlot}
\title{Plot the smooth covariate effect}
\usage{
binomRegMethModelPlot(BEM.obj, mfrow = NULL, same.range = FALSE)
}
\arguments{
\item{BEM.obj}{an output object from function \code{binomRegMethModel}}

\item{mfrow}{the plot parameters to specify the layout of each plot}

\item{same.range}{specify whether the plots should be in the same vertical scale}
}
\value{
This function prints out a plot of smooth covariate effects and
its pointwise confidence intervals
}
\description{
This function accepts an output object from function
\code{binomRegMethModel} and print out a plot of the estimated covariate effect
across the region for each test covariate.
}
\examples{
#------------------------------------------------------------#
head(RAdat)
RAdat.f <- na.omit(RAdat[RAdat$Total_Counts != 0, ])
out <- binomRegMethModel(
  data=RAdat.f, n.k=rep(5, 3), p0=0.003307034, p1=0.9,
  epsilon=10^(-6), epsilon.lambda=10^(-3), maxStep=200, detail=FALSE
)
binomRegMethModelPlot(out, same.range=FALSE)
}
\author{
Kaiqiong Zhao
}
