% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSMethSim.R
\name{binomRegMethModelSim}
\alias{binomRegMethModelSim}
\title{Simulate Bisulfite sequencing data from specified smooth covariate effects}
\usage{
binomRegMethModelSim(n, posit, theta.0, beta, random.eff = FALSE,
  mu.e = 0, sigma.ee = 1, p0 = 0.003, p1 = 0.9, X, Z,
  binom.link = "logit")
}
\arguments{
\item{n}{sample size}

\item{posit}{genomic position; a numeric vector of size \code{p} (the number of CpG sites in the considered region).}

\item{theta.0}{a functional parameter for the intercept of the GAMM model; a numeric vector of size \code{p}.}

\item{beta}{a functional parameter for the slope of cell type composition. a numeric vector of size \code{p}}

\item{random.eff}{indicate whether adding the subject-specific random effect term \code{e}.}

\item{mu.e}{the mean of the random effect; a single number.}

\item{sigma.ee}{variance of the random effect; a single positive number.}

\item{p0}{the probability of observing a methylated read when the underlying true status is unmethylated.}

\item{p1}{the probability of observing a methylated read when the underlying true status is methylated.}

\item{X}{the matrix of the read coverage for each CpG in each sample; a matrix of n rows and \code{p} columns}

\item{Z}{numeric vector of length p for the covariate; currently, the covariate is the percentage of Cell type A (considering that the samples are composed of two cell types A and B); (Added on Feb 2018), Z can be a matrix; we allow for more than one covariates}

\item{binom.link}{the link  function used for simulation}
}
\value{
The function returns a list of following objects
\itemize{
\item \code{S} the true methylation counts; a numeric matrix of \code{n} rows and \code{p} columns
\item \code{Y} the observed methylation counts; a numeric matrix of \code{n} rows and \code{p} columns
\item \code{theta} the methylation parameter (after the logit transformation); a numeric matrix of \code{n} rows and \code{p} columns
\item \code{pi}
}
}
\description{
Simulate Bisulfite sequencing data from a Generalized Additive Model with functional parameters varying with the genomic position. Both the true methylated counts and observed methylated counts are generated, given the error/conversion rate parameters \code{p0} and \code{p1}.
}
\author{
Kaiqiong Zhao
}
